Updated perform_fft_on_epochs and updated y_frequencies_dict Format
So far what I have done is I have reshaped the y _frequencies dict to be made completely of dicionaries. This way it is easier to make sure you are acessing the right data because the dictionary is so complex now. Furthermore, I have also updated the perform fft on epochs function becuase it was averaging in the time domain instead of first applying the fft to the epoch data and then averaging. So it was essentially doing something to finding the power spectrum of the ERP instead of doing what was in gyurkovics so I changed that.